Separate native_connector_l2_adapter.py into individual files
