+++ lmcache/v1/distributed/l2_adapters/native_connector_l2_adapter.py
@@ -22,16 +22,9 @@ from __future__ import annotations
 
 # Standard
 from collections import defaultdict
-from typing import TYPE_CHECKING, Any, Optional
-import importlib
 import os
 import select
 import threading
-
-if TYPE_CHECKING:
-    from lmcache.v1.distributed.internal_api import (
-        L1MemoryDesc,
-    )
 
 # First Party
 from lmcache.logging import init_logger
@@ -40,13 +33,6 @@ from lmcache.v1.distributed.api import ObjectKey
 from lmcache.v1.distributed.l2_adapters.base import (
     L2AdapterInterface,
     L2TaskId,
-)
-from lmcache.v1.distributed.l2_adapters.config import (
-    L2AdapterConfigBase,
-    register_l2_adapter_type,
-)
-from lmcache.v1.distributed.l2_adapters.factory import (
-    register_l2_adapter_factory,
 )
 from lmcache.v1.memory_management import MemoryObj
 
@@ -357,371 +343,3 @@ class NativeConnectorL2Adapter(L2AdapterInterface):
                         self._completed_loads[task_id] = bitmap
                         os.eventfd_write(self._load_efd, 1)
 
-
-# -------------------------------------------------------------------
-# Config class
-# -------------------------------------------------------------------
-
-
-class RESPL2AdapterConfig(L2AdapterConfigBase):
-    """
-    Config for an L2 adapter backed by a native RESP
-    connector (Redis/Valkey).
-
-    Fields:
-    - host: server hostname or IP.
-    - port: server port.
-    - num_workers: C++ worker threads for I/O (default 8).
-    - username: optional auth username.
-    - password: optional auth password.
-    """
-
-    def __init__(
-        self,
-        host: str,
-        port: int,
-        num_workers: int = 8,
-        username: str = "",
-        password: str = "",
-    ):
-        super().__init__()
-        self.host = host
-        self.port = port
-        self.num_workers = num_workers
-        self.username = username
-        self.password = password
-
-    @classmethod
-    def from_dict(cls, d: dict) -> "RESPL2AdapterConfig":
-        host = d.get("host")
-        if not isinstance(host, str) or not host:
-            raise ValueError("host must be a non-empty string")
-
-        port = d.get("port")
-        if not isinstance(port, int) or port <= 0:
-            raise ValueError("port must be a positive integer")
-
-        num_workers = d.get("num_workers", 8)
-        if not isinstance(num_workers, int) or num_workers <= 0:
-            raise ValueError("num_workers must be a positive integer")
-
-        username = d.get("username", "")
-        password = d.get("password", "")
-
-        return cls(
-            host=host,
-            port=port,
-            num_workers=num_workers,
-            username=str(username),
-            password=str(password),
-        )
-
-    @classmethod
-    def help(cls) -> str:
-        return (
-            "RESP L2 adapter config fields:\n"
-            "- host (str): Redis/Valkey server hostname "
-            "or IP (required)\n"
-            "- port (int): server port (required, >0)\n"
-            "- num_workers (int): C++ worker threads "
-            "for I/O (default 8, >0)\n"
-            "- username (str): auth username "
-            "(default empty)\n"
-            "- password (str): auth password "
-            "(default empty)"
-        )
-
-
-def _create_resp_l2_adapter(
-    config: L2AdapterConfigBase,
-    l1_memory_desc: "Optional[L1MemoryDesc]" = None,
-) -> L2AdapterInterface:
-    """Create a NativeConnectorL2Adapter backed by the
-    C++ Redis connector."""
-    try:
-        # First Party
-        from lmcache.lmcache_redis import (
-            LMCacheRedisClient,
-        )
-    except ImportError as e:
-        raise RuntimeError(
-            "RESP L2 adapter requires the C++ Redis "
-            "extension. Build with: pip install -e ."
-        ) from e
-
-    assert isinstance(config, RESPL2AdapterConfig)
-    native_client = LMCacheRedisClient(
-        config.host,
-        config.port,
-        config.num_workers,
-        config.username,
-        config.password,
-    )
-    logger.info(
-        "Created RESP L2 adapter: %s:%d (workers=%d)",
-        config.host,
-        config.port,
-        config.num_workers,
-    )
-    return NativeConnectorL2Adapter(native_client)
-
-
-# Self-register config type and adapter factory
-register_l2_adapter_type("resp", RESPL2AdapterConfig)
-register_l2_adapter_factory("resp", _create_resp_l2_adapter)
-
-
-# -------------------------------------------------------------------
-# FS native L2 adapter config + factory
-# -------------------------------------------------------------------
-
-
-class FSNativeL2AdapterConfig(L2AdapterConfigBase):
-    """
-    Config for an L2 adapter backed by the native C++
-    filesystem connector.
-
-    Fields:
-    - base_path: directory for storing KV cache files.
-    - num_workers: C++ worker threads for I/O (default 4).
-    - relative_tmp_dir: relative sub-dir for temp files.
-    - use_odirect: bypass page cache via O_DIRECT.
-    - read_ahead_size: trigger filesystem readahead by
-      reading this many bytes first (optional).
-    """
-
-    def __init__(
-        self,
-        base_path: str,
-        num_workers: int = 4,
-        relative_tmp_dir: str = "",
-        use_odirect: bool = False,
-        read_ahead_size: Optional[int] = None,
-    ):
-        self.base_path = base_path
-        self.num_workers = num_workers
-        self.relative_tmp_dir = relative_tmp_dir
-        self.use_odirect = use_odirect
-        self.read_ahead_size = read_ahead_size
-
-    @classmethod
-    def from_dict(cls, d: dict) -> "FSNativeL2AdapterConfig":
-        base_path = d.get("base_path")
-        if not isinstance(base_path, str) or not base_path:
-            raise ValueError("base_path must be a non-empty string")
-
-        num_workers = d.get("num_workers", 4)
-        if not isinstance(num_workers, int) or num_workers <= 0:
-            raise ValueError("num_workers must be a positive integer")
-
-        relative_tmp_dir = d.get("relative_tmp_dir", "")
-        if not isinstance(relative_tmp_dir, str):
-            raise ValueError("relative_tmp_dir must be a string")
-
-        use_odirect = d.get("use_odirect", False)
-        if not isinstance(use_odirect, bool):
-            raise ValueError("use_odirect must be a boolean")
-
-        read_ahead_size = d.get("read_ahead_size", None)
-        if read_ahead_size is not None:
-            if not isinstance(read_ahead_size, int) or read_ahead_size <= 0:
-                raise ValueError("read_ahead_size must be a positive integer")
-
-        return cls(
-            base_path=base_path,
-            num_workers=num_workers,
-            relative_tmp_dir=str(relative_tmp_dir),
-            use_odirect=use_odirect,
-            read_ahead_size=read_ahead_size,
-        )
-
-    @classmethod
-    def help(cls) -> str:
-        return (
-            "FS native L2 adapter config fields:\n"
-            "- base_path (str): directory for KV "
-            "cache files (required)\n"
-            "- num_workers (int): C++ worker threads "
-            "for I/O (default 4, >0)\n"
-            "- relative_tmp_dir (str): relative "
-            "sub-dir for temp files (default empty)\n"
-            "- use_odirect (bool): bypass page cache "
-            "via O_DIRECT (default false)\n"
-            "- read_ahead_size (int): trigger fs "
-            "readahead by reading this many bytes "
-            "first (optional)"
-        )
-
-
-def _create_fs_native_l2_adapter(
-    config: L2AdapterConfigBase,
-    l1_memory_desc: "Optional[L1MemoryDesc]" = None,
-) -> L2AdapterInterface:
-    """Create a NativeConnectorL2Adapter backed by the
-    C++ filesystem connector."""
-    try:
-        # First Party
-        from lmcache.lmcache_fs import (
-            LMCacheFSClient,
-        )
-    except ImportError as e:
-        raise RuntimeError(
-            "FS native L2 adapter requires the C++ FS "
-            "extension. Build with: pip install -e ."
-        ) from e
-
-    assert isinstance(config, FSNativeL2AdapterConfig)
-    native_client = LMCacheFSClient(
-        config.base_path,
-        config.num_workers,
-        config.relative_tmp_dir,
-        config.use_odirect,
-        config.read_ahead_size or 0,
-    )
-    logger.info(
-        "Created FS native L2 adapter: %s (workers=%d, odirect=%s, read_ahead=%s)",
-        config.base_path,
-        config.num_workers,
-        config.use_odirect,
-        config.read_ahead_size,
-    )
-    return NativeConnectorL2Adapter(native_client)
-
-
-register_l2_adapter_type("fs_native", FSNativeL2AdapterConfig)
-register_l2_adapter_factory("fs_native", _create_fs_native_l2_adapter)
-
-
-# -------------------------------------------------------------------
-# Native plugin L2 adapter: dynamically load a third-party
-# pybind-wrapped C++ connector and wrap it with
-# NativeConnectorL2Adapter.
-# -------------------------------------------------------------------
-
-
-class NativePluginL2AdapterConfig(L2AdapterConfigBase):
-    """Config for a native-plugin L2 adapter.
-
-    Dynamically loads a third-party pybind-wrapped C++
-    connector class and wraps it with
-    ``NativeConnectorL2Adapter``.
-
-    Fields:
-    - module_path: Dotted Python import path of the
-        module containing the connector class.
-    - class_name: Name of the connector class inside
-        *module_path*.
-    - adapter_params: Arbitrary dict forwarded to the
-        connector class constructor as keyword arguments.
-    """
-
-    def __init__(
-        self,
-        module_path: str,
-        class_name: str,
-        adapter_params: dict[str, Any] | None = None,
-    ):
-        self.module_path = module_path
-        self.class_name = class_name
-        self.adapter_params = adapter_params or {}
-
-    @classmethod
-    def from_dict(cls, d: dict) -> "NativePluginL2AdapterConfig":
-        module_path = d.get("module_path")
-        if not isinstance(module_path, str) or not module_path:
-            raise ValueError("module_path must be a non-empty string")
-
-        class_name = d.get("class_name")
-        if not isinstance(class_name, str) or not class_name:
-            raise ValueError("class_name must be a non-empty string")
-
-        adapter_params = d.get("adapter_params", {})
-        if not isinstance(adapter_params, dict):
-            raise ValueError("adapter_params must be a dict")
-
-        return cls(
-            module_path=module_path,
-            class_name=class_name,
-            adapter_params=adapter_params,
-        )
-
-    @classmethod
-    def help(cls) -> str:
-        return (
-            "Native plugin L2 adapter config fields:\n"
-            "- module_path (str): dotted import path of "
-            "the module containing the connector "
-            "class (required)\n"
-            "- class_name (str): name of the connector "
-            "class inside the module (required)\n"
-            "- adapter_params (dict): forwarded as "
-            "kwargs to the connector constructor "
-            "(optional, default {})\n"
-            "\n"
-            "Example JSON:\n"
-            '{"type": "native_plugin", '
-            '"module_path": "my_ext.connector", '
-            '"class_name": "MyConnectorClient", '
-            '"adapter_params": '
-            '{"host": "localhost", "port": 1234}}'
-        )
-
-
-def _create_native_plugin_l2_adapter(
-    config: L2AdapterConfigBase,
-    l1_memory_desc: "Optional[L1MemoryDesc]" = None,
-) -> L2AdapterInterface:
-    """Dynamically load a third-party native connector
-    and wrap it with NativeConnectorL2Adapter."""
-    assert isinstance(config, NativePluginL2AdapterConfig)
-
-    try:
-        module = importlib.import_module(config.module_path)
-    except ImportError as e:
-        raise ImportError(
-            "Could not import module '%s': %s" % (config.module_path, e)
-        ) from e
-
-    try:
-        connector_cls = getattr(module, config.class_name)
-    except AttributeError as e:
-        raise AttributeError(
-            "Module '%s' has no class '%s': %s"
-            % (config.module_path, config.class_name, e)
-        ) from e
-
-    native_client = connector_cls(**config.adapter_params)
-
-    # Verify the native client exposes the required
-    # interface methods.
-    required_methods = [
-        "event_fd",
-        "submit_batch_get",
-        "submit_batch_set",
-        "submit_batch_exists",
-        "drain_completions",
-        "close",
-    ]
-    for method in required_methods:
-        if not callable(getattr(native_client, method, None)):
-            raise TypeError(
-                "%s.%s instance missing required method "
-                "'%s'"
-                % (
-                    config.module_path,
-                    config.class_name,
-                    method,
-                )
-            )
-
-    logger.info(
-        "Created native plugin L2 adapter: %s.%s (params=%s)",
-        config.module_path,
-        config.class_name,
-        config.adapter_params,
-    )
-    return NativeConnectorL2Adapter(native_client)
-
-
-register_l2_adapter_type("native_plugin", NativePluginL2AdapterConfig)
-register_l2_adapter_factory("native_plugin", _create_native_plugin_l2_adapter)
